server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
python-multipart==0.0.6
mangum==0.17.0
PyJWT==2.8.0
python-dotenv==1.1.1
supabase==1.3.0
pydantic[email]==2.5.0